requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
msal>=1.24.0
openpyxl>=3.1.0
openai>=1.0.0
anthropic>=0.18.0

